requirements.txt
streamlit==1.38.0
pandas==2.2.2
python-dateutil==2.9.0
matplotlib==3.8.4